Shuffle around cover files, and restructure how the machine<->cover attachment interaction occurs
