+++ src/main/java/com/gregtechceu/gtceu/api/machine/MachineDefinition.java
@@ -119,9 +119,6 @@ public class MachineDefinition implements Supplier<IMachineBlock> {
     @Getter
     @Setter
     private Supplier<BlockState> appearance;
-    @Getter
-    @Setter
-    private boolean allowCoverOnFront;
     @Nullable
     @Getter
     @Setter
