+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/MultiblockMachineBuilder.java
@@ -89,7 +89,6 @@ public class MultiblockMachineBuilder extends MachineBuilder<MultiblockMachineDe
         super(registrate, name, MultiblockMachineDefinition::new, metaMachine::apply, blockFactory,
                 itemFactory, blockEntityFactory);
         allowExtendedFacing(true);
-        allowCoverOnFront(true);
         // always add the formed property to multi controllers
         modelProperty(IMultiController.IS_FORMED_PROPERTY, false);
     }
@@ -474,11 +473,6 @@ public class MultiblockMachineBuilder extends MachineBuilder<MultiblockMachineDe
     @Override
     public MultiblockMachineBuilder allowExtendedFacing(boolean allowExtendedFacing) {
         return (MultiblockMachineBuilder) super.allowExtendedFacing(allowExtendedFacing);
-    }
-
-    @Override
-    public MultiblockMachineBuilder allowCoverOnFront(boolean allowCoverOnFront) {
-        return (MultiblockMachineBuilder) super.allowCoverOnFront(allowCoverOnFront);
     }
 
     @Override
