+++ src/main/java/com/gregtechceu/gtceu/common/cover/io/PumpCover.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.common.cover;
+package com.gregtechceu.gtceu.common.cover.io;
 
 import com.gregtechceu.gtceu.api.GTValues;
 import com.gregtechceu.gtceu.api.capability.IControllable;
