+++ src/main/java/com/gregtechceu/gtceu/common/data/GTCovers.java
@@ -8,9 +8,14 @@ import com.gregtechceu.gtceu.api.addon.IGTAddon;
 import com.gregtechceu.gtceu.api.cover.CoverDefinition;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
 import com.gregtechceu.gtceu.client.renderer.cover.*;
-import com.gregtechceu.gtceu.common.cover.*;
 import com.gregtechceu.gtceu.common.cover.detector.*;
 import com.gregtechceu.gtceu.common.cover.ender.EnderFluidLinkCover;
+import com.gregtechceu.gtceu.common.cover.filter.FluidFilterCover;
+import com.gregtechceu.gtceu.common.cover.filter.ItemFilterCover;
+import com.gregtechceu.gtceu.common.cover.io.*;
+import com.gregtechceu.gtceu.common.cover.misc.ComputerMonitorCover;
+import com.gregtechceu.gtceu.common.cover.misc.FacadeCover;
+import com.gregtechceu.gtceu.common.cover.misc.MachineControllerCover;
 import com.gregtechceu.gtceu.common.cover.voiding.AdvancedFluidVoidingCover;
 import com.gregtechceu.gtceu.common.cover.voiding.AdvancedItemVoidingCover;
 import com.gregtechceu.gtceu.common.cover.voiding.FluidVoidingCover;
