+++ src/main/java/com/gregtechceu/gtceu/common/data/machines/GTMachineUtils.java
@@ -216,8 +216,7 @@ public class GTMachineUtils {
                             .rotationState(RotationState.ALL)
                             .colorOverlayTieredHullModel(ioOverlay, pipeOverlay, emissiveOverlay)
                             .abilities(abilities)
-                            .tooltips(Component.translatable("gtceu.machine." + tooltip + ".tooltip"))
-                            .allowCoverOnFront(true);
+                            .tooltips(Component.translatable("gtceu.machine." + tooltip + ".tooltip"));
 
                     if (slots == 1) {
                         builder.tooltips(Component.translatable("gtceu.universal.tooltip.fluid_storage_capacity",
